ARD page creation
