+++ src/main/resources/ValidationMessages.properties
@@ -32,3 +32,4 @@ approval.selectionNotMade=Tell us which officer signed off the accounts
 offBalanceSheetArrangementsQuestion.selectionNotMade=Tell us if your prepared accounts include any ''Off balance sheet arrangements'' notes
 offBalanceSheetArrangements.details.missing=Give details of your ''off balance sheet arrangements''
 accountsReferenceDateQuestion.selectionNotMade=Please confirm the dates that you have prepared your accounts to
+accountsReferenceDate.selectionNotMade=Please confirm your new company account date
